PyAutoGUI changed to Requests
Ported the main code from pyautogui to requests, so now you don't have to stay on the same discord channel/discord window just for it to work. Due to this it also now requires a token. (Can be obtained from the "Authorization" header on a request to the discord api on the inspect element network tab)